+++ core/components/romanescobackyard/migxconfigs/content_related.config.js
@@ -21,7 +21,7 @@
                         "allowBlank": "1",
                         "depth": "0",
                         "limitRelatedContext": "false",
-                        "where": "[{\"deleted:=\":\"0\"},{\"context_key:!=\":\"hub\"}]",
+                        "where": "[{\"deleted:=\":\"0\"},{\"context_key:NOT IN\":[\"[[++romanesco.link_exclude_contexts:replace=` ==`:replace=`,==\",\"`]]\"]}]",
                         "pageSize": "20"
                     },
                     "restrictive_condition": "",
@@ -48,7 +48,7 @@
                         "allowBlank": "1",
                         "depth": "0",
                         "limitRelatedContext": "false",
-                        "where": "[{\"deleted:=\":\"0\"},{\"context_key:!=\":\"hub\"}]",
+                        "where": "[{\"deleted:=\":\"0\"},{\"context_key:NOT IN\":[\"[[++romanesco.link_exclude_contexts:replace=` ==`:replace=`,==\",\"`]]\"]}]",
                         "pageSize": "20"
                     },
                     "restrictive_condition": "",
@@ -94,6 +94,27 @@
                     "sourceFrom": "config",
                     "sources": "",
                     "inputOptionValues": "",
+                    "default": "",
+                    "useDefaultIfEmpty": "0",
+                    "pos": ""
+                },
+                {
+                    "MIGX_id": "",
+                    "field": "category",
+                    "caption": "Category",
+                    "description": "",
+                    "description_is_code": "0",
+                    "inputTV": "",
+                    "inputTVtype": "listbox",
+                    "validation": "",
+                    "configs": {
+                        "typeAhead": "false"
+                    },
+                    "restrictive_condition": "",
+                    "display": "",
+                    "sourceFrom": "config",
+                    "sources": "",
+                    "inputOptionValues": "@CHUNK tvSelectInputOption@LinkCategory",
                     "default": "",
                     "useDefaultIfEmpty": "0",
                     "pos": ""
@@ -319,9 +340,24 @@
         },
         {
             "MIGX_id": "",
+            "header": "Category",
+            "dataIndex": "category",
+            "width": 50,
+            "sortable": true,
+            "show_in_grid": 1,
+            "customrenderer": "",
+            "renderer": "",
+            "clickaction": "",
+            "selectorconfig": "",
+            "renderchunktpl": "",
+            "renderoptions": "",
+            "editor": ""
+        },
+        {
+            "MIGX_id": "",
             "header": "Weight",
             "dataIndex": "weight",
-            "width": 20,
+            "width": 30,
             "sortable": true,
             "show_in_grid": 1,
             "customrenderer": "",